+++ pkg/reconciliation/reconcile_services.go
@@ -99,9 +99,6 @@ func (rc *ReconciliationContext) CheckHeadlessServices() result.ReconcileResult 
 			// if we found the service already, check if they need updating
 			if !utils.ResourcesHaveSameHash(currentService, desiredSvc) {
 				resourceVersion := currentService.GetResourceVersion()
-				// preserve any labels and annotations that were added to the service post-creation
-				desiredSvc.Labels = utils.MergeMap(map[string]string{}, currentService.Labels, desiredSvc.Labels)
-				desiredSvc.Annotations = utils.MergeMap(map[string]string{}, currentService.Annotations, desiredSvc.Annotations)
 
 				// ClusterIP may have been updated for the NodePort service
 				// so we need to preserve it.  Copying should not break any of